+++ src/main/java/lotto/exceptions/InvalidLottoNumberException.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.exceptions;
 
 public class InvalidLottoNumberException extends RuntimeException {
 	public InvalidLottoNumberException() {
