+++ src/main/java/lotto/exceptions/InvalidLottoNumberLengthException.java
@@ -1,7 +1,7 @@
-package lotto;
+package lotto.exceptions;
 
 public class InvalidLottoNumberLengthException extends RuntimeException {
-	InvalidLottoNumberLengthException() {
+	public InvalidLottoNumberLengthException() {
 		super("로또번호는 총 6개여야 합니다");
 	}
 }
